requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-anthropic>=0.0.1
openai>=1.0.0
anthropic>=0.5.0
python-dotenv>=1.0.0
tenacity>=8.2.0
pydantic>=2.7.0,<3.0.0
typing-extensions>=4.5.0
boto3>=1.28.0
pandas>=2.0.0
pytest>=7.0.0
moto>=4.0.0
aws-cdk-lib>=2.120.0
constructs>=10.3.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation>=0.42b0
mypy>=1.7.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
python-magic>=0.4.27
graphviz>=0.20.1
pydot>=1.4.2
markdown>=3.5.1
jinja2>=3.1.2
flask==3.0.2
flask-cors==4.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pyarrow==14.0.1
openpyxl==3.1.2
fastavro==1.8.2
httpx==0.25.2
langchain-core>=0.3.58 